+++ src/main/perl/lib/t/01-credentials.t
@@ -1,7 +1,7 @@
 use strict;
 use warnings;
 
-use Test::More tests => 4;
+use Test::More tests => 6;
 
 use File::Temp qw/:mktemp/;
 use File::Path;
@@ -17,6 +17,9 @@ my $credentials_file = eval {
   
   open (my $fh, ">$home/.aws/credentials") or BAIL_OUT("could not create temporary credentials file");
   print $fh <<eot;
+[default]
+region = us-west-1
+
 [foo]
 aws_access_key_id=foo-aws-access-key-id
 aws_secret_access_key=foo-aws-secret-access-key
@@ -24,6 +27,8 @@ aws_secret_access_key=foo-aws-secret-access-key
 [bar]
 aws_access_key_id=bar-aws-access-key-id
 aws_secret_access_key=bar-aws-secret-access-key
+region = us-east-1
+
 eot
   close $fh;
   "$home/.aws/credentials";
@@ -33,9 +38,11 @@ $ENV{HOME} = "$home";
 my $creds = new Amazon::Credentials({ order => [qw/file/] });
 ok(ref($creds), 'find credentials');
 is($creds->get_aws_access_key_id, 'foo-aws-access-key-id', 'default profile');
+is($creds->get_region, 'us-west-1', 'default region');
 
 $creds = new Amazon::Credentials({ profile => 'bar', order => [qw/file/] });
 is($creds->{aws_access_key_id}, 'bar-aws-access-key-id', 'retrieve profile');
+is($creds->get_region, 'us-east-1', 'region');
 
 END {
   eval {